requirements.txt
torch==2.7.1
torchvision==0.22.1
numpy==2.2.5
scipy==1.15.3
pandas==2.3.0
scikit-learn==1.7.0
scikit-image==0.25.2
opencv-python==4.11.0.86
pillow==11.3.0
accelerate==1.8.1
pyyaml==6.0.2
tqdm==4.67.1